+++ app/admin/courses.rb
@@ -50,7 +50,25 @@ ActiveAdmin.register Course do
   filter :description
   filter :start_booking_hours
   filter :end_booking_minutes
-  filter :state       , as: :select, collection: Course.localized_states
+  filter :state           , as: :select, collection: Course.localized_states
   filter :created_at
   filter :updated_at
+
+  form do |f|
+    f.inputs do
+      f.semantic_errors *f.object.errors.keys
+      if current_admin_user.is_root?
+        f.input :organization
+      else
+        f.input :organization, collection: [current_admin_user.organization]
+      end
+      f.input :category, collection: current_admin_user.categories
+      f.input :name
+      f.input :description
+      f.input :start_booking_hours
+      f.input :end_booking_minutes
+      f.input :state
+    end
+    f.actions
+  end
 end